应用的 executor都托管给 ShutdownExecutorServicesOnUninstallEventHandler
