+++ koupleless-base-plugin/src/main/java/com/alipay/sofa/koupleless/plugin/concurrent/KouplelessExecutorService.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.common.api;
+package com.alipay.sofa.koupleless.plugin.concurrent;
+
+import com.alipay.sofa.koupleless.plugin.manager.handler.ShutdownExecutorServicesOnUninstallEventHandler;
 
 import java.util.Collection;
 import java.util.List;
@@ -35,6 +37,7 @@ public class KouplelessExecutorService implements ExecutorService {
 
     public KouplelessExecutorService(ExecutorService executorService) {
         this.executorService = executorService;
+        ShutdownExecutorServicesOnUninstallEventHandler.manageExecutorService(this.executorService);
     }
 
     @Override
