+++ koupleless-base-plugin/src/main/java/com/alipay/sofa/koupleless/plugin/concurrent/KouplelessScheduledExecutorService.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.common.api;
+package com.alipay.sofa.koupleless.plugin.concurrent;
+
+import com.alipay.sofa.koupleless.plugin.manager.handler.ShutdownExecutorServicesOnUninstallEventHandler;
 
 import java.util.Collection;
 import java.util.List;
@@ -36,6 +38,8 @@ public class KouplelessScheduledExecutorService implements ScheduledExecutorServ
 
     public KouplelessScheduledExecutorService(ScheduledExecutorService scheduledExecutorService) {
         this.scheduledExecutorService = scheduledExecutorService;
+        ShutdownExecutorServicesOnUninstallEventHandler
+            .manageExecutorService(this.scheduledExecutorService);
     }
 
     @Override
