+++ koupleless-base-plugin/src/test/java/com/alipay/sofa/koupleless/plugin/concurrent/KouplelessExecutorServiceTest.java
@@ -14,33 +14,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.common.api;
+package com.alipay.sofa.koupleless.plugin.concurrent;
 
 import com.alipay.sofa.ark.common.util.ClassLoaderUtils;
+import com.google.common.collect.Lists;
 import org.junit.After;
 import org.junit.Test;
 
 import java.net.URLClassLoader;
 import java.util.concurrent.Callable;
-import java.util.concurrent.SynchronousQueue;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import static org.junit.Assert.assertEquals;
 
 /**
  * @author lianglipeng.llp@alibaba-inc.com
- * @version $Id: KouplelessThreadPoolExecutorTest.java, v 0.1 2024年05月10日 12:04 立蓬 Exp $
+ * @version $Id: KouplelessExecutorServiceAdaptorTest.java, v 0.1 2024年05月14日 00:05 立蓬 Exp $
  */
-public class KouplelessThreadPoolExecutorTest {
+public class KouplelessExecutorServiceTest {
+    private ExecutorService executorService = Executors.newFixedThreadPool(6);
 
-    private URLClassLoader               classLoader = mockClassLoader();
-    private KouplelessThreadPoolExecutor executor    = new KouplelessThreadPoolExecutor(10, 10, 60L,
-        TimeUnit.SECONDS, new SynchronousQueue<Runnable>());
+    private ExecutorService executor        = new KouplelessExecutorService(executorService);
+
+    private URLClassLoader  classLoader     = mockClassLoader();
 
     @Test
     public void test() {
         ClassLoader currentCl = Thread.currentThread().getContextClassLoader();
         try {
+            Thread.currentThread().setContextClassLoader(classLoader);
 
             Runnable runnable = () -> assertEquals(classLoader,
                 Thread.currentThread().getContextClassLoader());
@@ -50,11 +56,16 @@ public class KouplelessThreadPoolExecutorTest {
                 return "mock";
             };
 
-            Thread.currentThread().setContextClassLoader(classLoader);
             executor.submit(runnable);
             executor.submit(callable);
             executor.submit(runnable, "mock");
+            executor.invokeAll(Lists.newArrayList(callable));
+            executor.invokeAll(Lists.newArrayList(callable), 2, TimeUnit.SECONDS);
+            executor.invokeAny(Lists.newArrayList(callable));
+            executor.invokeAny(Lists.newArrayList(callable), 2, TimeUnit.SECONDS);
 
+        } catch (InterruptedException | ExecutionException | TimeoutException e) {
+            throw new RuntimeException(e);
         } finally {
             Thread.currentThread().setContextClassLoader(currentCl);
         }
